fix: 整理小说埋点代码规范
